death-ghost-reclaim cycle
git-svn-id: http://localhost:88/svn/uwow4/uwow2/trunk@525 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ src/RealmServer.Core/Objects/Update/UpdateObjectDtoMapper.cs
@@ -254,7 +254,7 @@ namespace Hazzik.Objects.Update {
 			//obj.Coinage = uint.MaxValue;
 			UpdateUnit(dto, obj);
 			dto.Set(UpdateFields.PLAYER_DUEL_ARBITER, obj.DuelArbiterGuid);
-			dto.Set(UpdateFields.PLAYER_FLAGS, obj.Flags);
+			dto.Set(UpdateFields.PLAYER_FLAGS, (uint)obj.Flags);
 			dto.Set(UpdateFields.PLAYER_GUILDID, obj.GuildId);
 			dto.Set(UpdateFields.PLAYER_GUILDRANK, obj.GuildRank);
 			dto.Set(UpdateFields.PLAYER_BYTES,
